Update qLib.shelf
First the first, would like to congratulate you and give you a big THANK YOU! for this tools, they are simply amazing and they make our lifes better each day at work.

This feature tested on CentOS Hou 16.5.439 qLib Toggle Integer Frames qL was not working properly on my side overall that is missing the set frame inc so actually was still displaying frame in step of 1 which was not creating so much difference.

I slightly modified and use it in this way and hope could be usefull from your side or if you have any comment or improvement to add. For sure, and please. Feel free to do it.

Thanks and wish a wonderful one! Cheers

Leo